requirements_minimal.txt
kiteconnect>=4.0.0
gspread>=5.0.0
google-auth>=2.0.0
pandas>=1.5.0
numpy>=1.24.0
pandas_ta>=0.3.0
requests>=2.28.0
# transformers>=4.30.0  # COMMENTED OUT FOR NOW